webui: mv scroll behavior to sketch-timeline
Moves the automatic scrolling behavior from sketch-app-shell into
sketch-timeline, where it makes more sense.

This change also makes automatic scrolling conditional on
an internal "scrollingState", which we update whenever you
manually scroll to, or away from the bottom of the timeline.

If you scroll to the bottom of the timeline, then it's "sticky"
and newly arriving messages will keep scrolling you to the bottom
as they render.

If you scroll up to older messages though, then we stop automatically
scrolling you to the latest messages at the bottom of the timeline.

When the timeline is in this latter "floating" scrollingState, we
also render a floating down-arrow button over the lower right corner
of the timeline.  Clicking on this will take you down to the latest
message at the end of the timeline.