sketch: diff commenting: avoid re-sending old messages
When I was adding diff comments, it was sending me back to the chat
screen, which isn't what I wanted, since I want to send multiple
comments in one go. Furthermore, it was re-sending old comments
because of some new-to-me attribute stuff.

Finally I moved the comment box out of its right-hand corner to
the line I'm commenting on. (Sketch did this part.)

Co-Authored-By: sketch