Sketch "inception" support.
I wanted to tell Sketch-in-container to run

	go run ./cmd/sketch -unsafe -skaband-addr= -one-shot 'what is todays date'"

and have it work. And now it does.

The underlying complexity is whether "inDocker" is enough to know
"isInsideSketch", so I separated that as far as necessary. The skaband
stuff is still complicated.

It might be preferable to use env variables for all outside->inside
communication rather than use flags sometimes and variables sometimes
(like, for keys), but here we are.