dockerimg: set http.postBuffer for git inside container
We sometimes see failures with git push from the innie to the outie.
It appears to be when there is a big binary in the commit
(which is usually because of a different bug).
The internet suggests the HTTP POST limit in the git client
may be an issue (1MB in the git version in debian bookworm),
so let's bump it up.