dockerimg: remove tmp-sketch-binary-linux
This is a remnant from when we tried to include the sketch
binary in the container image we built. Instead of doing
that, we `docker cp` the sketch binary into the container
every time we run it. (This is to make sure that sketch is
always fresh when we are developing it, and do not find
ourselves stuck on the docker image cache.)

Because of the docker cp step, there is no longer any
need to give the binary a funny name and include it in the
directory we build the container from. So stop doing that.

A bonus is that sketch will stop accidentally adding it
to commits.

Fixes #39