dockerimg: use --progress=quiet to always show docker build failure
In non-verbose mode, --progress=quiet is very quiet when everything
works, but prints an actual error when something fails (e.g. the
dockerfile is bad or you ran out of space).

While here, remove the partial stdout/stderr plumbing that did not
quite work (because the code sometimes used os.Stdout/os.Stderr
directly).

For #37
For #41