dockerimg: more Dockerfile robustness
LLM-generated RUN commands such as

	RUN . .venv/bin/activate && uv pip sync requirements.txt || true

were failing when .venv didn't exist, because under the default
debian shell (dash), sourcing a non-existant file causes a command
to fail even with `|| true`.

Make things a little easier for the LLM by switching to bash.
Additionally, configure bash with set +e for the duration of the
LLM-generated extra_cmds.