loop: update git commit instruction to use --trailer, add Change-ID
We have run 'git config' for the user, so we can trim those instructions and simplify.
(And in unsafe mode, the user has 'git config' set up anyway.)

Instead of prompting the model to manually add 'Co-Authored-By' as a line in the
commit message, now instruct it to use the --trailer option.
This streamlines adding a Change-ID trailer with a random string (s<random_hex>k format).

I'd actually like to use precommmit hooks to automatically do all the trailer,
but that doesn't play nicely with -unsafe, so for now, do it this way.
At least we'll have Change-IDs that we can start using.
(Pity the official change-id support in git hasn't landed yet.)

Co-Authored-By: sketch <hello@sketch.dev>
Change-ID: s9b68cbfa4c7eeb82k