agent: move "sketch-base" into git
The agent's notion of "initial commit" is kind of special, in that it
is used as the "base" for a bunch of git operations. It's hard for
the user to change (we only provide a workflow via restart), yet
sometimes you want to do just that.

So, instead we put it as data inside of it, named as a tag sketch-base.
It's abusing tags, but branches are no better.