sketch main: migrating things from /init into cmdline flags
As much as possible, I want NewAgent() to take as many arguments as it
can, and try to avoid doing different things with Init().

You can't quite do everything because the port that Docker has
open for forwarding starts as "0" and we call "docker port" to find it,
but besides that, almost everything is knowable up front.