sketch: avoid using tty when termui is false.
When using with -termui=false -verbose, the terminal does weird shit,
and there's a ladder effect. It's because a tty is being allocated
somehow. We don't need a tty unless we're using termui, so tightening up
the if/else on when we pass "-t" to Docker.

The logic here is detailed, so apologies if I got it wrong.