llm: get costs from server
Calculating costs on the client has the advantage
that it works when not using skaband.

It requires that we maintain multiple sources of truth, though.
And it makes it very challenging to add serverside tools,
such as Anthropic's web tool.

This commit switches sketch to rely on the server for all costs.
If not using skaband, no costs will be calculated, which also
means that budget constraints won't work.

It's unfortunate, but at the moment it seems like the best path.