all: s/title/slug/, adjust branch handling
There are two intertwined changes here.

First, replace title with slug, and precommit with commit-message-style.

The slug makes enough of a title, and it provides a single human-readable
identifier we can use everywhere.

Second, construct the branch name on the fly instead of storing it,
out of slug, branch prefix, and retryNumber.
This removes some duplicated data, and makes the retry loop
easier to follow and reason about.