Fix up sketch-wip handling when --commit is not specified.
If -git-remote-url and -commit are both specified, we check out
sketch-wip. We should do so always when we're in a container, so
doing that now.

The new sketch-wip stuff is weird if we're not in a container (the
"unsafe" mode without -outside-hostname). I've changed the system
prompt slightly to try to figure that out.