Using GH actions provided Chrome for test speedup (#149)
OMG. This had wild behavior. Sometimes you'd log into tmate, and it
would work. Something about running Chrome manually would tickle some
dbus thing into working. The error would sometimes show up, but most
often it would be swallowed.

The punchline? We had a context timeout of 5 or 10 seconds in the tests
that would trigger silently, and we'd see nothing. Everything else
(including maybe the dbus stuff) was a red herring: the test executor
machine is swamped, and shit's slow, and the various timeouts needed
tweaking. I bet there are a few more timeouts to loosen eventually, but
it seems to be passing now.

Error message:
	failed to start browser (please apt get chromium or equivalent): chrome failed to start: