webui: fix Monaco fonts
In Chrome, if you selectected a line in the diff view, it wouldn't
select the whole line, because Monaco was somehow confused about the
font widths or something. Turns out we had customized our fonts. We
don't need to do that!

Also adding a sketchDebug idea so that it's easier to poke at things.