Re-enable monaco horizontal scroll bars.
Like all things LLM and Monaco, the key here is to delete things
that aren't default until things get better. The scrollbars don't
show up because they aren't necessary, and, when the horizontal
ones are necessary, they now show up.

I'm still seeing two bugs:

1. If you click into the diff the first time on some line, especially
if it's the second diff block, Monaco jumps to a different area. This
is kind of terrible and annoying.

2. My touchpad should let me scroll right/left as well as vertical, but
I'm only getting vertical scrolling in this mode.