webui: improve comment functionality for Monaco diff editor by using gutter
This was a bit of a journey. To me, the correct functionality is like
the breakpoint setting in VSCode. I had o3 find the relevant code and
tried to feed it back into Sketch, but it turned out I had to break it
down into pieces. (Asking it for the whole thing failed several times.)
In the end, I asked Sketch first to add a comment icon in the gutters.
This required just the right "glyphMargin" settings and finally worked.
In the next session, it worked on showing/hiding them depending on where
you were hovering. In the one after that, it deleted the old comment box
functionality, and then in the one after that, added that functionality
back in. And now we're here, and it seems to work.

Co-Authored-By: sketch <hello@sketch.dev>
Change-ID: sd81ff0507f3398a8k