webui: improve/simplify commit selector in Monaco diff view
The HTML select's cant be formatted to show tags nicely, so
did a "custom" select thing.

Furthermore, I never use the "to" thing, so let's just get rid of
it, and we're always going from someplace to someplace.

I don't love the overflow behavior, but we'll see how it goes.

Co-Authored-By: sketch <hello@sketch.dev>
Change-ID: s9f70e9aef1fb08c4k