server: Wait on terminal process for cleanup
The bash shell for the termianl is an exec.Cmd that is Start()ed
in the github.com/creack/pty package. We do a lot of cleanup
when that shell exits, but we never Wait() on it (and neither
does the pty package), so the bash process gets left behind
as a zombie.

Fixes boldsoftware/sketch#181