sketch: create empty commit on empty repo, both in innie and outtie
Life's a bit too short to fail when the git repo exists but is empty.
For innie (e.g., using -unsafe on an empty repo), we just create the
empty commit. For outtie, we do the same thing, since initializing an
empty repo would be a weird different code path.

Co-Authored-By: sketch <hello@sketch.dev>
Change-ID: se0a9606cc9f43388k