Revert "cmd/sketch: add a process reaper"
The reaper was too fast, and it was reaping processes
that we started with os/exec before os/exec could Wait
on them, which caused Wait to return an error.

Most of our error handling for Wait causes bad things to happen:
sketch exits, diffs look scary, etc.

Revert now, and I'll resubmit later when it's more robust.

This reverts commit 587b57dce0d2301e5628ed3d39e62f20d58d7f48.