sketch: replace inline zombie cleanup with periodic reaper
This is a partial revert of sorts of the bashtool smorgasborg.
I couldn't reliably repro, but some daemons would fail to start
entirely, and not totally reliably. There was certainly a race
somewhere, and this is the most likely thing.

Remove process group killing and inline zombie reaping from cmdWait.
Add periodic zombie reaper that runs every minute and reaps zombies
that have been in that state for 5+ minutes. Only runs when PID 1.

Co-Authored-By: sketch <hello@sketch.dev>
Change-ID: sd6673f56fa6a3560k